Now the popout menu work and changes the colors of the items, but need to work on the 'reset color' one as it makes the image invisible
